Add delete button back
