+++ client/src/components/Modals/Resource/styles.scss
@@ -51,4 +51,22 @@
       opacity: 0.9;
     }
   }
+
+  .general-checkbox {
+    margin-top: 20px !important;
+  }
+
+  .delete-btn .button {
+    background-color: white !important;
+    color: #575757 !important;
+    border: solid 1px #575757 !important;
+    float: left;
+    font-weight: bold;
+    margin: 20px 20px 20px 0px;
+
+    &:hover {
+      background-color: transparent !important;
+      opacity: 0.7 !important;
+    }
+  }
 }
